% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmatXL.R
\name{Arow1}
\alias{Arow1}
\title{@title Row 1 of matrix \strong{A}}
\usage{
Arow1(ped3)
}
\arguments{
\item{ped3}{: A pedigree \code{data.frame} with columns corresponding to ID, SIRE, DAM.
This is usually a fraction of a pedigree extracted from an ancestor, which is the output of \code{peddown}.}
}
\value{
A.row1 : Genetic relationships with an ancestor; a \code{data.frame} with columns corresponding to ID, rg.
}
\description{
Calculates the 1st row of a genetic relationship matrix.
}
