% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A-tubular.R
\name{buildA}
\alias{buildA}
\title{Relationship matrix \strong{A}}
\usage{
buildA(ped)
}
\arguments{
\item{ped}{: The pedigree \code{data.frame} with integer columns corresponding to ID, SIRE, DAM.}
}
\value{
A : Relationship matrix \strong{A}
}
\description{
Creates the pedigree-based additive genetic relationship matrix.
}
\examples{
ped = data.frame(ID=1:6, SIRE=c(0,0,1,3,1,4), DAM=c(0,0,2,2,2,5))
buildA(ped)

}
