% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildD.R
\name{buildD}
\alias{buildD}
\title{Relationship matrix \strong{D}}
\usage{
buildD(ped, A)
}
\arguments{
\item{ped}{: \code{data.frame} with integer columns corresponding to ID, SIRE, DAM. Missing value is 0.}

\item{A}{: Relationship matrix \strong{A} created by function \code{\link{buildA}}.}
}
\value{
Relationship \code{matrix} \strong{D}
}
\description{
Builds the pedigree-based dominance relationship matrix.
}
\examples{
ped = data.frame(ID=1:6, SIRE=c(0,0,1,3,1,4), DAM=c(0,0,2,2,2,5))
buildD(ped, buildA(ped))

}
