% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inb.R
\name{inb}
\alias{inb}
\title{Inbreeding coefficient}
\usage{
inb(ped, id)
}
\arguments{
\item{ped}{: \code{data.frame} with integer columns corresponding to ID, SIRE, DAM. Missing value is 0.}

\item{id}{: Numeric ID of an individual}
}
\value{
Inb : Inbreeding coefficient of the individual
}
\description{
Calculates inbreeding coefficient for an individual.
}
\examples{
ped = data.frame(ID=1:7, SIRE=c(0,0,1,1,3,1,5), DAM=c(0,0,0,2,4,4,6))
inb(ped, 7)

}
