% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbreed.R
\name{inbreed}
\alias{inbreed}
\title{Inbreeding coefficients}
\usage{
inbreed(ped)
}
\arguments{
\item{ped}{: \code{data.frame} with integer columns corresponding to ID, SIRE, DAM. Missing value is 0.}
}
\value{
Vector of inbreeding coefficients
}
\description{
Calculates inbreeding coefficients for all animals in the pedigree.
}
\examples{
ped = data.frame(ID=1:7, SIRE=c(0,0,1,1,3,1,5), DAM=c(0,0,0,2,4,4,6))
inbreed(ped)

}
