% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2tab.R
\name{mat2tab}
\alias{mat2tab}
\title{Matrix to tabular}
\usage{
mat2tab(mat)
}
\arguments{
\item{mat}{: \code{matrix}}
}
\value{
tab : \code{data.frame} with 2 integer (IDs) and 1 numeric (values) columns.
}
\description{
Converts matrix data to tabular data.
}
\examples{
ped = data.frame(ID=1:6, SIRE=c(0,0,1,3,1,4), DAM=c(0,0,2,2,2,5))
mat2tab(buildA(ped))

}
