% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmatXL.R
\name{peddown}
\alias{peddown}
\title{Pedigree extraction from an ancestor}
\usage{
peddown(ped2, indv)
}
\arguments{
\item{ped2}{: The output of \code{gghead}; for more details: \code{?gghead}
indv : An integer ID of an ancestor}
}
\value{
newped : An extracted pedigree \code{data.frame} with columns corresponding to ID, SIRE, DAM.
}
\description{
Extracts pedigree from an ancestor.
}
