% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedup.R
\name{pedup}
\alias{pedup}
\title{Upward pedigree extraction}
\usage{
pedup(ped, progeny)
}
\arguments{
\item{ped}{: \code{data.frame} with integer columns corresponding to ID, SIRE, DAM. Missing value is 0.}

\item{progeny}{: Vector of animal(s), from which the new pedigree is being extracted.}
}
\value{
newped : Extracted pedigree \code{data.frame}
}
\description{
Extract pedigree upward for one or a group of animals to find their ascendants
}
\examples{
ped = data.frame(ID=1:6, SIRE=c(0,0,1,3,1,4), DAM=c(0,0,2,2,2,5))
pedup(ped, c(1,4))

}
