% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qgpu.R
\name{Qgpu}
\alias{Qgpu}
\title{Vector \strong{Qg} + \strong{u}}
\usage{
Qgpu(Q, sol)
}
\arguments{
\item{Q}{: The output matrix from \code{qmat} (for more details: \code{?qmat})}

\item{sol}{: \code{data.frame} with 2 numeric columns corresponding to ID, EBV ([\strong{ĝ}, \strong{û}]), where \strong{ĝ} and \strong{û} are the genetic group and genetic merit solutions, respectively. The order of solutions must be the order of columns and the order of rows in matrix \strong{Q}.}
}
\value{
Vector of \strong{Qĝ} + \strong{û}
}
\description{
Adds genetic group contributions to the genetic merit of animals.
}
\examples{
ped = data.frame(ID=c(3,4,6,5), SIRE=c(1,3,4,1), DAM=c(2,2,5,2))
Q = qmat(gghead(ped))
ghat = c(0.1, -0.2)
uhat = seq(-1.5, 1.5, 1)
sol = data.frame(ID=1:6, EBV=c(ghat, uhat))
Qgpu(Q, sol)

}
