% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmat.R
\name{qmat}
\alias{qmat}
\title{Create matrix \strong{Q}}
\usage{
qmat(ped2)
}
\arguments{
\item{ped2}{: The output of \code{gghead}; for more details: \code{?gghead}}
}
\value{
Qmat : Matrix \strong{Q}
}
\description{
Creates the genetic group contribution matrix.
}
\examples{
qmat(gghead(sampleped))

}
