% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmatXL.R
\name{qmatXL}
\alias{qmatXL}
\title{Matrix \strong{Q} for large pedigrees}
\usage{
qmatXL(ped2)
}
\arguments{
\item{ped2}{: The output of \code{gghead}; for more details: \code{?gghead}}
}
\value{
Q : Matrix \strong{Q}
}
\description{
Creates a genetic group contribution matrix for large pedigrees.
}
\details{
Calculation of a genetic group contribution matrix for large pedigrees requires a lot of memory and time.
This might not be possible on ordinary computers, using the function \code{qmat}.
The function \code{qmatXL} takes less RAM and time, making the calculation possible on ordinary computers.
}
\examples{
ped = data.frame(ID=c(3,4,6,5), SIRE=c(1,3,4,1), DAM=c(2,2,5,2))
ped2 = gghead(ped)
qmatXL(ped2)

}
