% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg.R
\name{rg}
\alias{rg}
\title{Genetic relationship coefficient}
\usage{
rg(ped, id1, id2)
}
\arguments{
\item{ped}{: \code{data.frame} with integer columns corresponding to ID, SIRE, DAM. Missing value is 0.}

\item{id1}{: Numeric ID of an individual}

\item{id2}{: Numeric ID of an individual}
}
\value{
Genetic relationship coefficient between the two individuals
}
\description{
Calculates genetic relationship coefficient between two individuals.
}
\examples{
ped = data.frame(ID=1:7, SIRE=c(0,0,1,1,3,1,5), DAM=c(0,0,0,2,4,4,6))
rg(ped, 5, 6)

}
