% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleped.R
\docType{data}
\name{sampleped}
\alias{sampleped}
\title{Sample pedigree data}
\format{An object of class \code{data.frame} with 4 rows and 3 columns.}
\usage{
sampleped
}
\description{
Sample data for \code{ped}
}
\keyword{datasets}
