% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesol.R
\docType{data}
\name{samplesol}
\alias{samplesol}
\title{Sample solution data}
\format{An object of class \code{data.frame} with 6 rows and 2 columns.}
\usage{
samplesol
}
\description{
Sample data for \code{sol}
}
\keyword{datasets}
