% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab2mat.R
\name{tab2mat}
\alias{tab2mat}
\title{Tabular to matrix}
\usage{
tab2mat(tab)
}
\arguments{
\item{tab}{: \code{data.frame} with 2 integer (IDs) and 1 numeric (values) columns.}
}
\value{
Converted \code{data.frame} to \code{matrix}
}
\description{
Converts tabular data to matrix data.
}
\examples{
ped = data.frame(ID=1:6, SIRE=c(0,0,1,3,1,4), DAM=c(0,0,2,2,2,5))
tab2mat(tabA(ped))

}
