% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tub2mat.R
\name{tub2mat}
\alias{tub2mat}
\title{Tubular to matrix}
\usage{
tub2mat(tub)
}
\arguments{
\item{tub}{: \code{data.frame} with 2 integer (IDs) and 1 numeric (values) columns.}
}
\value{
\code{matrix}
}
\description{
Converts tubular data to matrix data.
}
\examples{
ped = data.frame(ID=1:6, SIRE=c(0,0,1,3,1,4), DAM=c(0,0,2,2,2,5))
tub2mat(tubA(ped))

}
