% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubAinv.R
\name{tubAinv}
\alias{tubAinv}
\title{Inverse of the relationship matrix \strong{A} in a tubular format}
\usage{
tubAinv(ped, inb)
}
\arguments{
\item{ped}{: \code{data.frame} with integer columns corresponding to ID, SIRE, DAM. Missing value is 0.}

\item{inb}{: Inbreeding coefficients in the order of animals in the relationship matrix. It can be derived from \code{buildA} or \code{tubularA}.}
}
\value{
Inverse of the genetic relationship \code{data.frame}
}
\description{
Creates the inverse of the pedigree-based additive genetic relationship matrix in a \code{data.frame}.
}
\details{
\code{inb = diag(buildA) - 1}, or

\code{inb = tubA(ped); inb = inb[inb[,1]==inb[,2],]$a - 1}
}
\examples{
ped = data.frame(ID=1:6, SIRE=c(0,0,1,3,1,4), DAM=c(0,0,2,2,2,5))
inb = c(0, 0, 0, 0.25, 0, 0.25)
tubAinv(ped, inb)

}
