% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghead.R
\name{gghead}
\alias{gghead}
\title{Append genetic groups to the pedigree}
\usage{
gghead(ped)
}
\arguments{
\item{ped}{: \code{data.frame} with integer columns corresponding to ID, SIRE, DAM. Missing value is 0.}
}
\value{
Processed pedigree \code{data.frame}
}
\description{
Appends genetic groups to the head of the pedigree and sorts it.
}
\details{
Consider this simple pedigree:

3 0 0

4 3 0

6 4 5

5 0 0

First, unknown parents are replaced with the corresponding genetic groups.

Please note that unknown parent IDs should be smaller than animal IDs.

3 1 2

4 3 2

6 4 5

5 1 2

Then, \code{gghead} is applied to this pedigree (see the example).
}
\examples{
ped = data.frame(ID=c(3,4,6,5), SIRE=c(1,3,4,1), DAM=c(2,2,5,2))
gghead(ped)

}
