% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qgpu.R
\name{Qgpu}
\alias{Qgpu}
\title{\strong{Qg} + \strong{u}}
\usage{
Qgpu(Qmat, sol)
}
\arguments{
\item{Qmat}{: The output matrix from function \code{qmat}}

\item{sol}{: A \code{data.frame} with 2 numeric columns corresponding to ID, EBV ([\strong{ĝ}, \strong{û}]).}
}
\value{
uhatplus : Vector of \strong{Qĝ} + \strong{û}
}
\description{
Add genetic group contributions to genetic merit of animals in a pedigree.
}
\examples{
Qmat = qmat(ped)
# For details about qmat and ped: ?qmat
sol = data.frame(ID=c(1:2,4:7), EBV=c(0.2,seq(-0.1,0.3,0.1)))
Qgpu(Qmat, sol)

}
