% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmat.R
\name{qmat}
\alias{qmat}
\title{Matrix \strong{Q}}
\usage{
qmat(ped2)
}
\arguments{
\item{ped2}{: The output \code{data.frame} from \code{\link{gghead}} (for more details: \code{?\link{gghead}})}
}
\value{
\strong{Q} \code{matrix}
}
\description{
Creates the genetic group contribution matrix.
}
\examples{
ped = data.frame(ID=c(3,4,6,5), SIRE=c(1,3,4,1), DAM=c(2,2,5,2))
ped2 = gghead(ped)
qmat(ped2)

}
